backend/requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
gunicorn==24.0.0
python-dotenv==1.2.1

# async + http
httpx==0.28.1
aiohttp==3.13.3

# scheduler
APScheduler==3.11.2

# supabase
supabase==2.28.3
postgrest==2.28.3
storage3==2.28.3

# web3 / blockchain
web3==7.14.1
eth-account==0.13.7
py-algorand-sdk==2.4.0
puyapy==5.8.1

# llm
google-generativeai==0.8.5

# utils
pydantic==2.12.4
orjson==3.11.4
pytz==2025.2
